sets up base flex manager class and integration with IterableApi
