+++ iterableapi-ui/src/main/java/com/iterable/iterableapi/ui/flex/IterableFlexViewViewModel.kt
@@ -1,12 +1,11 @@
 package com.iterable.iterableapi.ui.flex
 
-import org.json.JSONObject
-
-import android.view.View
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.iterable.iterableapi.*
+import com.iterable.iterableapi.IterableApi
+import com.iterable.iterableapi.IterableFlexManager
+import com.iterable.iterableapi.IterableFlexMessage
 
 class IterableFlexViewViewModel : ViewModel() {
 
@@ -15,47 +14,8 @@ class IterableFlexViewViewModel : ViewModel() {
         get() = _flexMessages
 
     init {
+        val messages = IterableApi.getInstance().getFlexManager().getMessages()
 
-        // example of a JSON payload being serialized
-        val flexMessageMetadata = FlexMessageMetadata(
-            "doibjo4590340oidiobnw",
-            "mbn8489b7ehycy",
-            "noj9iyjthfvhs",
-            false
-        )
-
-        val flexMessageDefaultAction = FlexMessageElementsDefaultAction(
-            "someType", "someAction"
-        )
-
-        val flexMessageButtons: List<FlexMessageElementsButton> = listOf(
-            FlexMessageElementsButton("reward-button", "REDEEM MEOW", "success")
-        )
-
-        val flexMessageText: List<FlexMessageElementsText> = listOf(
-            FlexMessageElementsText("body", "CATS RULE!!!", "label")
-        )
-
-        val flexMessageElements = FlexMessageElements(
-            "Iterable Coffee Shoppe",
-            "SAVE 15% OFF NOW",
-            "http://placekitten.com/200/300",
-            flexMessageDefaultAction,
-            flexMessageButtons,
-            flexMessageText
-        )
-
-        val payload = JSONObject()
-
-        val flexMessageJson = JSONObject()
-        flexMessageJson.put("metadata", flexMessageMetadata.toJSONObject())
-        flexMessageJson.put("elements", flexMessageElements.toJSONObject())
-        flexMessageJson.put("payload", payload)
-
-        _flexMessages.value = listOf(
-            IterableFlexMessage.fromJSONObject(flexMessageJson),
-            IterableFlexMessage.fromJSONObject(flexMessageJson),
-            IterableFlexMessage.fromJSONObject(flexMessageJson)
-        )
+        _flexMessages.value = messages
     }
 }
